+++ CHANGELOG.md
@@ -107,6 +107,7 @@
 - Rename Region\Type -> Unit\UnitType to make it search able and clear !2352 @KoeMai
 - Improve quality/tests of manual pickup RestAPI !2369 @KoeMai
 - Renmae Region\Type -> Unit\UnitType to make it search able and clear !2352 @KoeMai
+- Improve documentation tooling for OpenAPI RequestBody !2421 @KoeMai
 
 ## Dev/Test/CI stuff
 - Added Gitpod and refactored permissions !2386 !2399 !2402 !2428 @sefn @chriswalg